requirements.txt
Django==4.2.7
django_environ==0.11.2
djangorestframework==3.14.0
djangorestframework_simplejwt==5.3.0
exponent_server_sdk==2.0.0
sentry_sdk==1.36.0
celery[redis]==5.3.6
django-filter==23.5
django-debug-toolbar==3.2.2
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
django-redis==5.4.0
pillow==10.2.0
django-csp==3.7
graphene-django==3.2.0
django-graphql-jwt==0.4.0
dj-rest-auth==5.0.1